+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.14] - 2020-08-27
+### Added
+- support for "privacy" settings for `GeoData`
+- storage for special surface & way types for routes in `GeoDataExtra` class
+
 ## [0.9.13] - 2020-05-21
 ### Changed
 - better Java support for `GeoData`